26-05-17 무게별조회기능 수정 완벽구현 실패

+++ backend/src/main/java/com/ggori_salang/backend/Service/CafeService.java
@@ -31,7 +31,7 @@ public class CafeService {
     public List<CafeVO> searchCafes(
             int regionId,
             List<String> petTypes,
-            Double maxWeight
+            Integer maxWeight
     ) {
         List<CafeVO> cafes = cafeDAO.findByFilters(regionId, petTypes, maxWeight);
         attachImages(cafes);
